requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
asyncpg
python-decouple
